refactor: Language -> language
cleaning up the api while we still can...
