+++ benchmarks/src/test/scala/flatgraph/benchmark/CompileTests.scala
@@ -2,7 +2,7 @@ package flatgraph.benchmark
 
 import io.shiftleft.codepropertygraph.generated.*
 import io.shiftleft.codepropertygraph.generated.nodes.*
-import io.shiftleft.codepropertygraph.generated.Language.*
+import io.shiftleft.codepropertygraph.generated.language.*
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
