+++ test-schemas-domain-classes/src/main/scala/testdomains/generic/language.scala
@@ -1,13 +1,13 @@
-package testdomains.hierarchical
+package testdomains.generic
 
 /** combining all implicits into one trait that can be mixed in further downstream */
-trait Language
+trait language
     extends accessors.ConcreteStoredConversions
     with traversals.ConcreteStoredConversions
     with neighboraccessors.Conversions
-    with flatgraph.traversal.Language
+    with flatgraph.traversal.language
     with flatgraph.Implicits {
-  implicit def toGeneratedNodeStarters(domain: Hierarchical): HierarchicalNodeStarters = HierarchicalNodeStarters(domain)
+  implicit def toGeneratedNodeStarters(domain: GenericDomain): GenericDomainNodeStarters = GenericDomainNodeStarters(domain)
 }
 
-object Language extends Language
+object language extends language
