+++ test-schemas-domain-classes/src/main/scala/testdomains/generic/nodes/NodeB.scala
@@ -1,6 +1,6 @@
 package testdomains.generic.nodes
 
-import testdomains.generic.Language.*
+import testdomains.generic.language.*
 import scala.collection.immutable.{IndexedSeq, ArraySeq}
 
 /** Node base type for compiletime-only checks to improve type safety. EMT stands for: "erased marker trait", i.e. it is erased at runtime
