+++ test-schemas-domain-classes/src/main/scala/testdomains/gratefuldead/nodes/Artist.scala
@@ -1,6 +1,6 @@
 package testdomains.gratefuldead.nodes
 
-import testdomains.gratefuldead.Language.*
+import testdomains.gratefuldead.language.*
 import scala.collection.immutable.{IndexedSeq, ArraySeq}
 
 /** Node base type for compiletime-only checks to improve type safety. EMT stands for: "erased marker trait", i.e. it is erased at runtime
