+++ test-schemas-domain-classes/src/main/scala/testdomains/hierarchical/neighboraccessors/package.scala
@@ -1,6 +1,6 @@
 package testdomains.hierarchical
 
-import flatgraph.traversal.Language.*
+import flatgraph.traversal.language.*
 import testdomains.hierarchical.nodes
 
 package object neighboraccessors {
